+++ include/block.h
@@ -8,14 +8,14 @@ typedef struct blockdevice_s {
 	int blksz;
 	int (*rw)(int mode, int minor, u64 blk, char *buf);
 	int (*rw_multiple)(int mode, int minor, u64, char *buf, int);
-	int (*ioctl)(int min, int cmd, int arg);
+	int (*ioctl)(int min, int cmd, long arg);
 	int (*select)(int min, int rw);
 	unsigned char cache;
 	mutex_t acl;
 } blockdevice_t;
 void init_block_devs();
 blockdevice_t *set_blockdevice(int maj, int (*f)(int, int, u64, char*), 
-	int bs, int (*c)(int, int, int), int (*m)(int, int, u64, char *, int), int (*s)(int, int));
+	int bs, int (*c)(int, int, long), int (*m)(int, int, u64, char *, int), int (*s)(int, int));
 int block_rw(int rw, dev_t dev, u64 blk, char *buf, blockdevice_t *bd);
 
 
@@ -25,7 +25,7 @@ int block_device_rw(int mode, dev_t dev, off_t off, char *buf, size_t len);
 int block_ioctl(dev_t dev, int cmd, int arg);
 int do_block_rw(int rw, dev_t dev, u64 blk, char *buf, blockdevice_t *bd);
 int set_availablebd(int (*f)(int, int, u64, char*), int bs, 
-	int (*c)(int, int, int), int (*m)(int, int, u64, char *, int), int (*s)(int, int));
+	int (*c)(int, int, long), int (*m)(int, int, u64, char *, int), int (*s)(int, int));
 int do_block_rw_multiple(int rw, dev_t dev, u64 blk, char *buf, 
 	blockdevice_t *bd, int count);
 void block_cache_init();
